requirements.txt
fastapi==0.103.2
pydantic-settings==2.0.3
uvicorn[standard]==0.23.2
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pymysql==1.1.0
alembic==1.12.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
pytest==7.4.2
fastapi-mail==1.4.1
Jinja2==3.1.2
PyJWT==2.8.0